feat : dev pull 이전 commit

+++ src/main/java/com/newsfeed/domain/comment/service/CommentService.java
@@ -1,5 +1,6 @@
-package com.newsfeed.domain.comment;
+package com.newsfeed.domain.comment.service;
 
+import com.newsfeed.domain.auth.security.PrincipalDetails;
 import com.newsfeed.domain.comment.dto.*;
 import com.newsfeed.domain.comment.entity.Comment;
 
@@ -10,7 +11,6 @@ import com.newsfeed.domain.comment.repository.CommentRepository;
 import com.newsfeed.domain.newsfeed.entity.Newsfeed;
 import com.newsfeed.domain.newsfeed.repository.NewsfeedRepository;
 import com.newsfeed.domain.user.entity.User;
-import com.newsfeed.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,16 +24,14 @@ import java.util.List;
 public class CommentService {
     private final CommentRepository commentRepository;
     private final NewsfeedRepository newsfeedRepository;
-    private final UserRepository userRepository;
 
     //댓글 생성
-    public GlobalResponse<CommentCreateRes> create(int status, String message, CommentCreateReq req, Long newsfeedId, Long parentCommentId) {
+    public GlobalResponse<CommentCreateRes> create(int status, String message, CommentCreateReq req, Long newsfeedId, Long parentCommentId, PrincipalDetails principalDetails) {
 
         Newsfeed newsfeed = newsfeedRepository.findById(newsfeedId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.NEWSFEED_NOT_FOUND));
 
-        User user = userRepository.findById(req.getUserId())
-                .orElseThrow(() -> new NotFoundException(ErrorCode.USER_NOT_FOUND));
+        User user = principalDetails.getUser();
 
         //생성시에 부모가 없는 새 댓글일 수 있음. null, depth = 0 으로 새로 계층이 형성되는 댓글시에는 해당 변수 사용
         Comment parentComment = null;
@@ -62,18 +60,18 @@ public class CommentService {
     }
 
     //게시글에 대한 부모계층 댓글 조회
-    public GlobalResponse<List<CommentWithChildrenResponse>> getFromNewsfeedsComments(int status, String message,Long newsfeedId) {
+    public GlobalResponse<List<CommentFindResponse>> getFromNewsfeedsComments(int status, String message, Long newsfeedId) {
 
             Newsfeed newsfeed = newsfeedRepository.findById(newsfeedId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.NEWSFEED_NOT_FOUND));
 
             List<Comment> comments = commentRepository.findAll();
 
-            List<CommentWithChildrenResponse> data = new ArrayList<>();
+            List<CommentFindResponse> data = new ArrayList<>();
 
             for (Comment comment : comments) {
                 if (comment.getParentCommentId() == null && comment.getNewsfeedId().getId().equals(newsfeedId)) {
-                    data.add(new CommentWithChildrenResponse(comment));
+                    data.add(new CommentFindResponse(comment));
                 }
             }
 
@@ -81,26 +79,32 @@ public class CommentService {
     }
 
     //댓글에 대한 대댓글 조회
-    public GlobalResponse<List<CommentWithChildrenResponse>> getFromCommentsChildren(int status, String message, Long commentId) {
+    public GlobalResponse<List<CommentFindResponse>> getFromCommentsChildren(int status, String message, Long commentId) {
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new NotFoundException(ErrorCode.COMMENT_NOT_FOUND));
 
         List<Comment> children = commentRepository.findByParentCommentId_Id(comment.getId());
-        List<CommentWithChildrenResponse> data = new ArrayList<>();
+        List<CommentFindResponse> data = new ArrayList<>();
         for (Comment child : children) {
-            data.add(new CommentWithChildrenResponse(child));
+            data.add(new CommentFindResponse(child));
         }
 
         return GlobalResponse.success(status, message, data);
     }
 
     //댓글 수정
-    public GlobalResponse<CommentPutResponse> update(int status, String message,Long commentId ,CommentPutRequest req) {
+    public GlobalResponse<CommentPutResponse> update(int status, String message,Long commentId ,CommentPutRequest req, PrincipalDetails principalDetails) {
+
+        User user = principalDetails.getUser();
 
         Comment comment = commentRepository
                 .findById(commentId)
                 .orElseThrow(()-> new NotFoundException(ErrorCode.COMMENT_NOT_FOUND));
+
+        if (!comment.getUserId().getId().equals(user.getId())) {
+            throw new NotFoundException(ErrorCode.USER_NOT_MATCH);
+        }
 
         comment.update(req);
         commentRepository.save(comment);
@@ -109,10 +113,17 @@ public class CommentService {
     }
 
     //댓글 삭제
-    public GlobalResponse<Void> delete(int status, String message, Long commentId) {
+    public GlobalResponse<Void> delete(int status, String message, Long commentId, PrincipalDetails principalDetails) {
+
+        User user = principalDetails.getUser();
+
         Comment comment = commentRepository
                 .findById(commentId)
                 .orElseThrow(()-> new NotFoundException(ErrorCode.COMMENT_NOT_FOUND));
+
+        if (!comment.getUserId().getId().equals(user.getId())) {
+            throw new NotFoundException(ErrorCode.USER_NOT_MATCH);
+        }
 
         if (commentRepository.existsById(comment.getId())) {
             commentRepository.deleteById(comment.getId());
